Implemented Atomic structure Division of components into Atoms, molecules and organisms
