--- frontend/src/components/LoginModal.tsx
@@ -1,102 +1,0 @@
-import { useForm } from "react-hook-form";
-import { yupResolver } from "@hookform/resolvers/yup";
-import * as yup from "yup";
-import { User } from "../models/user";
-import { LoginCredentials } from "../network/notes_api";
-import * as NotesApi from "../network/notes_api";
-import { Alert, Button, Form, Modal } from "react-bootstrap";
-import TextInputField from "./form/TextInputField";
-import styleUtils from "../styles/utils.module.css";
-import { useState } from "react";
-import { UnauthorizedError } from "../errors/http_errors";
-import { useDispatch } from "react-redux";
-import { setUserInfo } from "../store/userSlice";
-
-interface LoginModalProps {
-  onDismiss: () => void;
-  onLoginSuccessful: (user: User) => void;
-}
-
-const LoginModal = ({ onDismiss, onLoginSuccessful }: LoginModalProps) => {
-  const [errorText, setErrorText] = useState<string | null>(null);
-
-  //yup schema for login credentials validation
-  const loginSchema = yup
-    .object({
-      username: yup.string().required("Username is required"),
-      password: yup.string().required("Password must be entered"),
-    })
-    .required();
-
-  const {
-    control,
-    register,
-    handleSubmit,
-    formState: { errors, isSubmitting, isValid, isDirty },
-  } = useForm<LoginCredentials>({
-    defaultValues: { username: "", password: "" },
-    resolver: yupResolver(loginSchema),
-    mode: "onChange",
-  });
-
-  const dispatch = useDispatch();
-
-  async function onSubmit(credentials: LoginCredentials) {
-    try {
-      const user = await NotesApi.login(credentials);
-      onLoginSuccessful(user);
-      dispatch(
-        setUserInfo({
-          username: user.username,
-          email: user.email,
-        })
-      );
-    } catch (error) {
-      if (error instanceof UnauthorizedError) {
-        setErrorText(error.message);
-      } else {
-        alert(error);
-      }
-      console.error(error);
-    }
-  }
-
-  return (
-    <Modal show onHide={onDismiss}>
-      <Modal.Header closeButton>
-        <Modal.Title>Log In</Modal.Title>
-      </Modal.Header>
-
-      <Modal.Body>
-        {errorText && <Alert variant="danger">{errorText}</Alert>}
-        <Form onSubmit={handleSubmit(onSubmit)}>
-          <TextInputField
-            name="username"
-            label="Username"
-            type="text"
-            placeholder="Username"
-            register={register}
-            error={errors.username}
-          />
-          <TextInputField
-            name="password"
-            label="Password"
-            type="password"
-            placeholder="Password"
-            register={register}
-            error={errors.password}
-          />
-          <Button
-            type="submit"
-            disabled={isSubmitting}
-            className={styleUtils.width100}
-          >
-            Log In
-          </Button>
-        </Form>
-      </Modal.Body>
-    </Modal>
-  );
-};
-
-export default LoginModal;
