--- frontend/src/components/NotesPageLoggedOutView.tsx
@@ -1,7 +1,0 @@
-const NotesPageLoggedOutView = () => {
-    return (
-        <p>Please login to see your notes</p>
-    );
-}
-
-export default NotesPageLoggedOutView;
