+++ frontend/src/components/atoms/index.ts
@@ -1,0 +1,4 @@
+export * from "./TextInputField";
+export * from "./verticalBarChart";
+export * from "./barChart";
+export * from "./doughnutChart"
