+++ frontend/src/components/atoms/verticalBarChart.tsx
@@ -18,7 +18,7 @@ ChartJS.register(
   Legend
 );
 
-export const options = {
+const options = {
   indexAxis: "y" as const,
   elements: {
     bar: {
@@ -39,7 +39,7 @@ export const options = {
 
 const labels = ["January", "February", "March", "April", "May", "June", "July"];
 
-export const data = {
+const data = {
   labels,
   datasets: [
     {
