+++ frontend/src/components/organisms/NavBar.tsx
@@ -1,7 +1,7 @@
 import { Container, Nav, Navbar } from "react-bootstrap";
-import { User } from "../models/user";
-import NavBarLoggedInView from "./NavBarLoggedInView";
-import NavBarLoggedOutView from "./NavBarLoggedOutView";
+import { User } from "../../models/user";
+import { NavBarLoggedInView } from "../molecules";
+import { NavBarLoggedOutView } from "../molecules";
 import { Link } from "react-router-dom";
 
 interface NavBarProps {
@@ -10,7 +10,7 @@ interface NavBarProps {
   onLoginClicked: () => void;
   onLogoutSuccessful: () => void;
 }
-const NavBar = ({
+export const NavBar = ({
   loggedInUser,
   onSignUpClicked,
   onLoginClicked,
@@ -44,5 +44,3 @@ const NavBar = ({
     </Navbar>
   );
 };
-
-export default NavBar;
