+++ frontend/src/components/organisms/NotesPageLoggedOutView.tsx
@@ -1,0 +1,3 @@
+export const NotesPageLoggedOutView = () => {
+  return <p>Please login to see your notes</p>;
+};
