+++ frontend/src/components/organisms/index.ts
@@ -1,0 +1,7 @@
+export * from "./LoginModal";
+export * from "./NavBar";
+export * from "./AddEditNoteDialog";
+export * from "./GraphMainSection";
+export * from "./NotesPageLoggedInView";
+export * from "./NotesPageLoggedOutView";
+export * from "./SignUpModal";
