+++ frontend/src/pages/GraphPage.tsx
@@ -1,6 +1,6 @@
 import { Container } from "react-bootstrap";
-import GraphMainSection from "../components/GraphMainSection";
-import NotesPageLoggedOutView from "../components/NotesPageLoggedOutView";
+import { GraphMainSection } from "../components/organisms";
+import {NotesPageLoggedOutView} from "../components/organisms";
 import { User } from "../models/user";
 import styles from "../styles/NotesPage.module.css";
 
@@ -8,7 +8,7 @@ interface NotesPageProps {
   loggedInUser: User | null;
 }
 //Parent component for Graph Screen, if user is not logged in shows public screen.
-const GraphPage = ({ loggedInUser }: NotesPageProps) => {
+export const GraphPage = ({ loggedInUser }: NotesPageProps) => {
   return (
     <Container className={styles.notesPage}>
       <>{loggedInUser ? <GraphMainSection /> : <NotesPageLoggedOutView />}</>
@@ -16,4 +16,3 @@ const GraphPage = ({ loggedInUser }: NotesPageProps) => {
   );
 };
 
-export default GraphPage;
