+++ frontend/src/pages/NotFoundPage.tsx
@@ -1,10 +1,8 @@
 //page for unauthorized users
-const NotFoundPage = () => {
+export const NotFoundPage = () => {
   return (
     <div>
       <p>Page not found</p>
     </div>
   );
 };
-
-export default NotFoundPage;
