+++ frontend/src/pages/NotesPage.tsx
@@ -1,24 +1,22 @@
 import { Container } from "react-bootstrap";
-import NotesPageLoggedInView from "../components/NotesPageLoggedInView";
-import NotesPageLoggedOutView from "../components/NotesPageLoggedOutView";
+import {
+  NotesPageLoggedOutView,
+  NotesPageLoggedInView,
+} from "../components/organisms";
 import { User } from "../models/user";
 import styles from "../styles/NotesPage.module.css";
 
 interface NotesPageProps {
-    loggedInUser: User | null,
+  loggedInUser: User | null;
 }
 //Parent component for Notes Screen, if user is not logged in shows public screen.
-const NotesPage = ({ loggedInUser }: NotesPageProps) => {
-    return (
-        <Container className={styles.notesPage}>
-            <>
-                {loggedInUser
-                    ? <NotesPageLoggedInView />
-                    : <NotesPageLoggedOutView />
-                }
-            </>
-        </Container>
-    );
-}
+export const NotesPage = ({ loggedInUser }: NotesPageProps) => {
+  return (
+    <Container className={styles.notesPage}>
+      <>
+        {loggedInUser ? <NotesPageLoggedInView /> : <NotesPageLoggedOutView />}
+      </>
+    </Container>
+  );
+};
 
-export default NotesPage;
