+++ frontend/src/pages/index.ts
@@ -1,0 +1,3 @@
+export * from "./GraphPage";
+export * from "./NotFoundPage";
+export * from "./NotesPage";
